Changed projects styling
