+++ src/App.tsx
@@ -1,11 +1,11 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from './components/Navbar/Navbar';
-import Home from './pages/Home';
-import Projects from './pages/Projects';
-import About from './pages/About';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import './styles/theme.css';
-import './styles/typography.css';
+import "bootstrap/dist/css/bootstrap.min.css";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
+import Navbar from "./components/Navbar/Navbar";
+import About from "./pages/About";
+import Home from "./pages/Home";
+import Projects from "./pages/Projects/Projects";
+import "./styles/theme.css";
+import "./styles/typography.css";
 
 function App() {
   return (
