+++ src/styles/typography.css
@@ -2,16 +2,14 @@ h1 {
   font-size: 3.5rem;
   font-weight: 700;
   margin-bottom: 1.5rem;
-  background: linear-gradient(120deg, var(--primary-light), var(--accent));
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
+  color: var(--neutral-dark);
   line-height: 1.2;
   letter-spacing: -0.02em;
 }
 
 h2 {
   font-size: 2.25rem;
-  color: var(--primary-lighter);
+  color: var(--primary-light);
   margin-bottom: 1rem;
   font-weight: 600;
   letter-spacing: -0.01em;
@@ -19,13 +17,13 @@ h2 {
 
 h3 {
   font-size: 1.75rem;
-  color: var(--accent);
+  color: var(--primary-light);
   font-weight: 500;
 }
 
 p {
   font-size: 1.1rem;
   line-height: 1.8;
-  color: var(--text-light);
+  color: var(--text-dark);
   opacity: 0.9;
 }
